feat: list 출력 부분을 컴포넌트화 시킴
